rename `rasa.test` [production code]
