+++ rasa/nlu/test.py
@@ -645,7 +645,7 @@ def _calculate_report(
     report_as_dict: Optional[bool] = None,
     exclude_label: Optional[Text] = None,
 ) -> Tuple[Union[Text, Dict], float, float, float, np.ndarray, List[Text]]:
-    from rasa.test import get_evaluation_metrics
+    from rasa.model_testing import get_evaluation_metrics
     import sklearn.metrics
     import sklearn.utils.multiclass
 
@@ -1930,7 +1930,7 @@ def _compute_metrics(
 
     Returns: metrics
     """
-    from rasa.test import get_evaluation_metrics
+    from rasa.model_testing import get_evaluation_metrics
 
     # compute fold metrics
     targets, predictions = _targets_predictions_from(
@@ -1952,7 +1952,7 @@ def _compute_entity_metrics(
 
     Returns: entity metrics
     """
-    from rasa.test import get_evaluation_metrics
+    from rasa.model_testing import get_evaluation_metrics
 
     entity_metric_results: EntityMetrics = defaultdict(lambda: defaultdict(list))
     extractors = get_entity_extractors(interpreter)
